requirements.txt
# This file was autogenerated by uv via the following command:
#    uv pip compile - -o requirements.txt
aiohappyeyeballs==2.6.1
    # via aiohttp
aiohttp==3.13.3
    # via discord-py
aiosignal==1.4.0
    # via aiohttp
attrs==25.4.0
    # via aiohttp
cffi==2.0.0
    # via pynacl
davey==0.1.4
discord-py==2.7.1
frozenlist==1.8.0
    # via
    #   aiohttp
    #   aiosignal
idna==3.11
    # via yarl
multidict==6.7.1
    # via
    #   aiohttp
    #   yarl
pathlib==1.0.1
propcache==0.4.1
    # via
    #   aiohttp
    #   yarl
pycparser==3.0
    # via cffi
pynacl==1.6.2
python-dotenv==1.2.2
typing-extensions==4.15.0
    # via aiosignal
yarl==1.23.0
    # via aiohttp
